random -> prevRandom rename to kilnV1
